requirements.txt
uvicorn[standard]
fastapi
sqlalchemy
python-dotenv
